+++ Leomard/Views/Popups/PostPopup.swift
@@ -69,16 +69,12 @@ struct PostPopup: View {
                                     alignment: .leading
                                 )
                                 .fontWeight(.semibold)
-                            HighlightedTextEditor(text: $commentText, highlightRules: .markdown)
-                                .overlay(RoundedRectangle(cornerRadius: 8).stroke(.primary, lineWidth: 0.5))
+                            MarkdownEditor(bodyText: $commentText, contentView: self.contentView)
                                 .frame(
                                     maxWidth: .infinity,
-                                    minHeight: 3 * NSFont.preferredFont(forTextStyle: .body).xHeight * 3,
-                                    maxHeight: .infinity,
-                                    alignment: .leading
+                                    minHeight: 3 * NSFont.preferredFont(forTextStyle: .body).xHeight,
+                                    maxHeight: .infinity
                                 )
-                                .lineLimit(5...)
-                                .font(.system(size: NSFont.preferredFont(forTextStyle: .body).pointSize))
                             Button("Send", action: createComment)
                                 .buttonStyle(.borderedProminent)
                                 .frame(
@@ -91,6 +87,7 @@ struct PostPopup: View {
                                     .progressViewStyle(.circular)
                             }
                         }
+                        .frame(maxHeight: .infinity)
                         Spacer()
                         Picker("Sort By", selection: $sortType) {
                             ForEach(CommentSortType.allCases, id: \.self) { method in
@@ -118,6 +115,7 @@ struct PostPopup: View {
                         }
                     }
                     .frame(
+                        maxWidth: .infinity,
                         minHeight: 0,
                         maxHeight: .infinity
                     )